requirements.txt
opencv-python-headless
keras
flask
sklearn
gunicorn
tensorflow-cpu==2.2.0
h5py == 2.10.0
scipy==1.4.1
Pillow